Things

+++ API Models/LolChampSelectChampSelectTimer.cs
@@ -4,7 +4,7 @@ namespace LCU.NET.API_Models
     {
         public int adjustedTimeLeftInPhase { get; set; }
         public int adjustedTimeLeftInPhaseInSec { get; set; }
-        public int internalNowInEpochMs { get; set; }
+        public long internalNowInEpochMs { get; set; }
         public bool isInfinite { get; set; }
         public string phase { get; set; }
         public int timeLeftInPhase { get; set; }
